requirements.txt
python-dotenv==0.20.0
requests==2.28.1
schedule==1.1.0
openai==1.6.1
instagrapi==2.0.1
Pillow>=8.1.1